refactor rename duplication and props dril
