+++ packages/files-ui/src/Components/Modules/FileBrowsers/views/FileSystemItem/SharedFolderRow.tsx
@@ -19,10 +19,10 @@ import { desktopSharedGridSettings, mobileSharedGridSettings } from "../../Share
 import SharedUsers from "../../../../Elements/SharedUser"
 import { t } from "@lingui/macro"
 import { Form, FormikProvider, useFormik } from "formik"
-import { object, string } from "yup"
 import clsx from "clsx"
 import { BucketKeyPermission } from "../../../../../Contexts/FilesContext"
 import UserBubble from "../../../../Elements/UserBubble"
+import { renameSchema } from "../../../../../Utils/validationSchema"
 
 const useStyles = makeStyles(({ breakpoints, constants, palette }: CSFTheme) => {
 
@@ -117,18 +117,6 @@ interface Props {
   handleRename: (bucket: BucketKeyPermission, newName: string) => void
 }
 
-const renameSchema = object().shape({
-  fileName: string()
-    .min(1, t`Please enter a name`)
-    .max(65, t`Name too long`)
-    .test(
-      t`Invalid name`,
-      t`Name cannot contain '/' character`,
-      (val: string | null | undefined) => !!val && !val?.includes("/")
-    )
-    .required(t`A name is required`)
-})
-
 const SharedFolderRow = ({ bucket, onFolderClick, menuItems, isEditing, setIsEditing, handleRename }: Props) => {
   const classes = useStyles()
   const { name, size } = bucket
@@ -203,6 +191,7 @@ const SharedFolderRow = ({ bucket, onFolderClick, menuItems, isEditing, setIsEdi
                 }}
                 placeholder = {t`Please enter a folder name`}
                 autoFocus={isEditing}
+                defaultValue={name}
               />
               <Button
                 data-cy='rename-submit-button'
